Removall of dependency to yaml and use introspection for c_ast in place of generation.
The goal is to have no generation of file even integrated.
python is enough concise to avoid usage of a dsl like YAML.

Introduction of introspection to act as the previously generated code.
In such way it is possible to enrichen the class of the AST and drift
from without the fear of alteration of the class in the future.

sub class Node creation through introspection through the function
subClassNode in ast which allow to have a more concise description
of the nodes in the c_ast module.