Add support for #pragma
Preprocessor pragmas and their arguments are tokenized (as PPPRAGMA and
PPPRAGMASTR) and included in the AST as a pppragma directive with the
argument as value. If no argument was given the string will be empty.

Unit test of the lexer, parser and generator have been modified and
added accordingly.

The previous behavior, that #pragma lines would be ignored, is
henceforth obsolete.