Add "get_context_parser()" to create 'sub'-Parser
this returns another instance of CParser that is created with a copy of
the parser's current known symbols.  This allows continuing parsing
in a given "context", e.g. one or more loaded header files.

There is also a "reset_identifiers" parameter on parse(), to accumulate
bits on the same parser.