Fix slow backtracking when checking for invalid strings
Use `regex` instead of `re`, and use atomic grouping.
To avoid surprises, use it everywhere.

https://pypi.org/project/regex/

> This regex implementation is backwards-compatible with the standard
> ‘re’ module, but offers additional functionality.

Fixes #61

This uses atomic grouping, which avoids the unnecessary backtracking.

> `(?>...)`
>
> If the following pattern subsequently fails, then the subpattern as a
whole will fail.

Also fix a test that relied on the incorrect handling of regexes.
The implementation documentation says that it intends to allow
**decimal** escapes permissively.

Install `regex` in Travis and Appveyor