Typedef for __kernel_sa_family_t in _fake_typedefs.h
This typedef that exists in the real `linux/socket.h` is masked by the
`fake_libc_include` hierarchy. Having it in place allows to successfully
parse a large number of Linux kernel headers necessary for netlink
operations. This is very useful for Python projects that communicate
with the kernel over netlink socket. In particular, this project:

https://github.com/ionos-cloud/netlinklib/tree/main/netlinklib

plans to use pyparser for parsing strcts defined in the kernel headers.

Signed-off-by: Eugene Crosser <crosser@average.org>